requirements.txt
torch
torchvision
timm
numpy
scikit-learn
pillow
pyyaml
tqdm
matplotlib
